Integrated module crawling from CORS, and other major modifications
NOTE: All templates and HTML files no longer point to relative CSS and JS files.

libs/
	-libs folder added, will contain any external python libraries for the application
	-BeautifulSoup 4 added as library

app.yaml
	-version changed to 0-2

css/fonts/
	-font files updated with new icons for share icons

css/main.css
	-added rule definitions for new share icons

css/skilltree.css
	-ensured that the control panel for each moduleBox div only appears in edit mode
	-adjusted the z-index of the moduleInfo and scroll bars

data.py
	-SaveSkillTreeHandler now can wipe a skill tree's information by setting an overwrite key value

modulecrawler.py
	-modulecrawler() renamed to crawl_modules()
	-urllib module replaced with urlfetch for performance
	-BeautifulSoup import modified
	-fixed inconsistency bugs while crawling

modulescraper.py
	-integrated modulecrawler module
	-decouple_multiple_module_codes() can now automatically create duplicate modules if they are not available in the provided module list.
	-fill_prerequisite_groups() typo fixed

not_found.html
	-added CSS imports and improved overall look

templates/SkillTree.html
	-left_panel now features an edit_panel and a view_panel, the default depending on whether the currently logged in user is the owner of the skill tree
	-skill tree is not editable at all if it is not in edit mode
	-disclaimerModal added, will display if the user is a new Student
	-shareModal added, will display share links if the share button is clicked.
	-now has logic to supply 2 GUIDs to the loadFromServer command if necessary.

skill_tree.py
	-now has logic to pass multiple skill trees and skill tree owners to the template

scripts/jquery.ajaxQueue
	-added library to allow automatic queueing of AJAX requests to prevent concurrent requests from blocking the browser rendering

scripts/script.js
	-added ability to filter according to the search input only after a 1 second delay
	-added ability to share the skill tree (shareLink()), and sharing after saving
	-added logic to overwrite the logged in user's skill tree when editing another's
	-ajaxQueue is now integrated
	-added jQuery functions massAddClass and massRemoveClass. Currently, these are monikers for standard addClass/removeClass until we can figure out how to use setTimeout efficiently.
	-added logic for the edit button.

under_construction.html
	-displays if the version on the online server is switched to under-construction

+++ data.py
@@ -239,9 +239,14 @@ class SaveSkillTreeHandler(webapp2.RequestHandler):
                 except AttributeError:
                     pass
                 logging.debug("Retrieved skill tree")
+                overwrite = urllib.unquote_plus(self.request.get('overwrite', default_value='false'))
+                logging.debug(overwrite)
+                if overwrite == 'true':
+                    for assigned_module in skill_tree.assigned_modules.iter(keys_only=True):
+                        assigned_module.delete()
+                    logging.debug("Skill tree cleared")
                 data = json.loads(data_string)
                 json_assigned_modules = data["assignedModules"]
-                json_remove_modules = data["removeModules"]
                 logging.debug("Parsed JSON data")
                 for json_assigned_module in json_assigned_modules.itervalues():
                     logging.debug("Inserting/updating module %s" % json_assigned_module["module"])
@@ -269,12 +274,14 @@ class SaveSkillTreeHandler(webapp2.RequestHandler):
                         assigned_module.prerequisites = json_assigned_module["prerequisites"]
                     assigned_module.put()
                     logging.debug("Inserted successfully")
-                for code, remove_module in json_remove_modules.iteritems():
-                    if remove_module:
-                        #module_k = datatypes.Module.query(datatypes.Module.code == code).get(keys_only=True)
-                        assigned_module_k = datatypes.AssignedModule.query(datatypes.AssignedModule.module_code == code).get(keys_only=True)
-                        assigned_module_k.delete()
-                        logging.debug("%s deleted." % code)
+                if hasattr(data, "removeModules"):
+                    json_remove_modules = data["removeModules"]
+                    for code, remove_module in json_remove_modules.iteritems():
+                        if remove_module:
+                            #module_k = datatypes.Module.query(datatypes.Module.code == code).get(keys_only=True)
+                            assigned_module_k = datatypes.AssignedModule.query(datatypes.AssignedModule.module_code == code).get(keys_only=True)
+                            assigned_module_k.delete()
+                            logging.debug("%s deleted." % code)
                 self.response.headers['Content-Type'] = 'text/plain'
                 self.response.write("SUCCESS")
     #def addOrUpdateModule(self, module, json_assigned_module):
